cosmetics fraud detect chatboat/requirements.txt
flask
pandas
google-generativeai
openpyxl
langchain-google-genai
langchain
python-dotenv